requirements.txt
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
PyPDF2==3.0.1
numpy==1.24.3